Add criticalCHRestartOccured
Website can indicate that a particular Client Hint is critical to the
page by including it in a `Critical-CH` HTTP response header. Doing so
will trigger a connection restart if the hint listed in the
`Critical-CH` HTTP response header could be (but wasn't) included in the
HTTP request initially sent. We should indicate this has happened in
Navigation Timing so that developers can know if a restart occurred and
impacted timing.

We can pull this from the [navigable](https://wicg.github.io/client-hints-infrastructure/#has-reloaded-for-critical-ch) but I'm not quite sure if this
is the right way to thread it.

Closes #177